@@ -1,89 +1,55 @@
-﻿using gymLog.API.Model;
-using gymLog.API.Model.DTO;
-using gymLog.API.Services.interfaces;
-using gymLog.Model;
-using BCrypt.Net;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.CodeAnalysis.Scripting;
-using Microsoft.IdentityModel.Tokens;
-using System;
-using System.IdentityModel.Tokens.Jwt;
+﻿using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
-using System.Threading.Tasks;
+using gymLog.API.Model;
+using gymLog.API.Model.DTO.AuthDto;
+using gymLog.API.Services.Interfaces;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 
-namespace gymLog.API.Controllers
+namespace gymLog.API.Controllers;
+
+[Route("api/[controller]")]
+[ApiController]
+public class AuthController : ControllerBase
 {
-    [Route("api/[controller]")]
-    [ApiController]
-    public class AuthController : ControllerBase
+    private readonly IAuthService _authService;
+    private readonly IJwtTokenGenerator _jwtTokenGenerator;
+
+    public AuthController(IAuthService authService, IJwtTokenGenerator jwtTokenGenerator)
     {
-        private readonly IUserService _userService;
-        private readonly IConfiguration _configuration;
+        _authService = authService;
+        _jwtTokenGenerator = jwtTokenGenerator;
+    }
 
-        public AuthController(IUserService userService, IConfiguration configuration)
+    [HttpPost("login")]
+    public async Task<IActionResult> Login([FromBody] LoginDto login)
+    {
+        var loginResult = await _authService.Login(login);
+        if (loginResult.IsSuccess)
         {
-            _userService = userService;
-            _configuration = configuration;
-        }
-
-        [HttpPost("login")]
-        public async Task<IActionResult> Login([FromBody] Login loginModel)
-        {
-            var user = await _userService.AuthenticateAsync(loginModel.Email, loginModel.PasswordHash);
-            if (user == null)
+            var token = _jwtTokenGenerator.GenerateToken(loginResult.Data);
+            return Ok(new
             {
-                return Unauthorized();
-            }
-
-            var token = GenerateJwtToken(user);
-            return Ok(new { token });
+                Token = token
+            });
         }
+        return BadRequest(loginResult.Message);
 
-        [HttpPost("register")]
-        public async Task<IActionResult> Register([FromBody] RegisterDto registerDto)
+    }
+
+    [HttpPost("register")]
+    public async Task<IActionResult> Register([FromBody] RegisterDto registerDto)
+    {
+        var registerResult = await _authService.Register(registerDto);
+        if (registerResult.IsSuccess)
         {
-            var user = new User
+            var token = _jwtTokenGenerator.GenerateToken(registerResult.Data);
+            return Ok(new
             {
-                Id = Guid.NewGuid(),
-                Name = registerDto.Name,
-                Email = registerDto.Email,
-                PasswordHash = BCrypt.Net.BCrypt.HashPassword(registerDto.Password),
-                DateOfBirth = registerDto.DateOfBirth,
-                Weight = registerDto.Weight,
-                Height = registerDto.Height,
-                CreatedAt = DateTime.UtcNow
-            };
-
-            var result = await _userService.CreateAsync(user);
-            if (result.IsSuccess) {
-                return CreatedAtAction(nameof(Login), new { email = result.Data!.Email }, result.Data);
-            }
-            
-            return BadRequest(result);
+                Token = token
+            });
         }
-
-        private string GenerateJwtToken(User user)
-        {
-            var jwtSettings = _configuration.GetSection("Jwt");
-            var key = Encoding.ASCII.GetBytes(jwtSettings["Key"]);
-
-            var tokenHandler = new JwtSecurityTokenHandler();
-            var tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Subject = new ClaimsIdentity(new[]
-                {
-                    new Claim(ClaimTypes.NameIdentifier, user.Id.ToString()),
-                    new Claim(ClaimTypes.Email, user.Email)
-                }),
-                Expires = DateTime.UtcNow.AddMinutes(double.Parse(jwtSettings["ExpireMinutes"])),
-                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature),
-                Issuer = jwtSettings["Issuer"],
-                Audience = jwtSettings["Audience"]
-            };
-
-            var token = tokenHandler.CreateToken(tokenDescriptor);
-            return tokenHandler.WriteToken(token);
-        }
+        return BadRequest(registerResult.Message);
     }
 }
