@@ -1,13 +1,13 @@
-﻿namespace gymLog.Model
+﻿namespace gymLog.API.Model;
+
+public class WorkoutPlan
 {
-    public class WorkoutPlan
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public ICollection<DayOfWeek> WorkoutSelectedDays { get; set; }
-        public ICollection<WorkoutDay> WorkoutDays { get; set; }
-        public Guid UserId { get; set; }
-        public User User { get; set; }
-    }
+    public Guid Id { get; set; }
+    public Guid UserId { get; set; }
+    public User User { get; set; } = null!;
+
+    public string Name { get; set; } = string.Empty;
+    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
+
+    public List<WorkoutDay> Days { get; set; } = new();
 }
