@@ -1,9 +1,14 @@
-﻿using gymLog.API.Services.interfaces;
-using gymLog.Model;
+﻿using gymLog.API.Model.DTO;
+using gymLog.API.Model.DTO.ExerciseDto;
+using gymLog.API.Model.DTO.WorkoutDayDto;
 
-namespace gymLog.API.Services
+namespace gymLog.API.Services.Interfaces;
+
+public interface IWorkoutDayService
 {
-    public interface IWorkoutDayService : IBasicCrudService<WorkoutDay>
-    {
-    }
+    public Task<Result<IEnumerable<WorkoutDayDto>>> GetAllWorkoutDays(Guid workoutPlanId);
+    public Task<Result<WorkoutDayDto>> GetWorkoutDayById(Guid workoutDayId);
+    public Task<Result<WorkoutDayDto>> CreateWorkoutDay(Guid workoutPlanId, CreateWorkoutDayDto createDayDto);
+    public Task<Result<WorkoutDayDto>> UpdateWorkoutDay(Guid workoutDayId, CreateWorkoutDayDto createDayDto);
+    public Task<Result<bool>> RemoveWorkoutDay(Guid workoutDayId);
 }
