@@ -1,18 +1,10 @@
-﻿using gymLog.API.Model;
+﻿namespace gymLog.API.Model;
 
-namespace gymLog.Model
+public class User
 {
-    public class User
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-        public string Email { get; set; }
-        public string PasswordHash { get; set; }
-        public DateTime DateOfBirth { get; set; }
-        public decimal Weight { get; set; }
-        public decimal Height { get; set; }
-        public DateTime CreatedAt { get; set; }
-        public ICollection<WorkoutPlan> WorkoutPlans { get; set; }
-        public ICollection<RefreshToken> RefreshTokens { get; set; }
-    }
+    public Guid Id { get; set; }
+    public string Email { get; set; } = string.Empty;
+    public string PasswordHash { get; set; } = string.Empty;
+    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
+    public List<WorkoutPlan> Plans { get; set; } = new();
 }
